Properly share helpers for creating Workspaces during unit tests
The only tests in AdhocWorkspaceTests that were using the helpers
really had no reason to be in AdhocWorkspaceTests so it's been moved
to a proper place.
